database: Not only updating hash if newer
Changed the database API to not change only the hash if we have
a newer file, but also the last_modified date.

Also testing the case if we try to change the hash of a
file that has the same age. Here, it is not clear what the
correct behavior should be.